+++ targetcxx/targetcxxcompiler.hpp
@@ -56,7 +56,7 @@ namespace Maike
 		private:
 			const TargetCxxOptions& r_options;
 			std::vector<const ParameterSet*> r_paramset;
-			mutable unsigned long int m_cxxversion_default;
+			mutable unsigned long long int m_cxxversion_default;
 
 			void execute(const Command& cmd,const char* source
 				,Twins<const FileInfo*> dependencies
